requirements.txt
numpy
pillow
flask
flask_cors
requests
torch
torchvision
gunicorn